back-end/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pypdfium2==5.7.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
reportlab==4.4.10
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
